requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
DateTime==5.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pytz==2023.3.post1
PyYAML==6.0.1
yarl==1.9.2
zope.interface==6.0
